Rewrite wabt::Opcode as C++ class (#500)
The class stores just one enum, so should be optimized to be passed as a
register. Since it is an object, the get_opcode_* accessors can now be
members:

    get_opcode_name(opcode) => opcode.GetName()

One side-effect of this change is that the opcode enumeration does not
match its binary encoding. This is a potential source of errors, but
will allows us to use a contiguous set of values for the enumeration, so
Opcode::GetInfo is an O(1) lookup. It will also be beneficial when
adding operators with prefixed encodings.

+++ src/binary-reader-opcnt.cc
@@ -77,10 +77,10 @@ BinaryReaderOpcnt::BinaryReaderOpcnt(OpcntData* data) : opcnt_data(data) {}
 
 Result BinaryReaderOpcnt::OnOpcode(Opcode opcode) {
   IntCounterVector& opcnt_vec = opcnt_data->opcode_vec;
-  while (static_cast<size_t>(opcode) >= opcnt_vec.size()) {
+  while (opcode >= opcnt_vec.size()) {
     opcnt_vec.emplace_back(opcnt_vec.size(), 0);
   }
-  ++opcnt_vec[static_cast<size_t>(opcode)].count;
+  ++opcnt_vec[opcode].count;
   return Result::Ok;
 }
